+++ extensions/jira/src/active-sprints.tsx
@@ -1,5 +1,5 @@
 import { List } from "@raycast/api";
-import { useCachedPromise } from "@raycast/utils";
+import { useCachedPromise, useCachedState } from "@raycast/utils";
 import { useState } from "react";
 
 import { getIssues } from "./api/issues";
@@ -15,12 +15,12 @@ export function ActiveSprints() {
     { keepPreviousData: true }
   );
 
-  const [projectKey, setProjectKey] = useState("");
+  const [projectKey, setProjectKey] = useCachedState("active-sprint-project", "");
   const jql = `sprint in openSprints() AND project = ${projectKey} ORDER BY updated DESC`;
 
   const {
     data: issues,
-    isLoading,
+    isLoading: isLoadingIssues,
     mutate,
   } = useCachedPromise((jql) => getIssues({ jql }), [jql], { execute: projectKey !== "" });
 
@@ -28,12 +28,12 @@ export function ActiveSprints() {
     <List.Dropdown
       tooltip="Filter issues by project"
       onChange={setProjectKey}
-      storeValue
+      value={projectKey}
       throttle
       isLoading={isLoadingProjects}
       onSearchTextChange={setProjectQuery}
     >
-      {projects?.map((project) => {
+      {projects.map((project) => {
         return (
           <List.Dropdown.Item
             key={project.id}
@@ -47,7 +47,12 @@ export function ActiveSprints() {
   ) : null;
 
   return (
-    <StatusIssueList issues={issues} isLoading={isLoading} mutate={mutate} searchBarAccessory={searchBarAccessory} />
+    <StatusIssueList
+      issues={issues}
+      isLoading={isLoadingIssues || isLoadingProjects}
+      mutate={mutate}
+      searchBarAccessory={searchBarAccessory}
+    />
   );
 }