+++ extensions/spotify-beta/src/components/TracksSection.tsx
@@ -6,24 +6,27 @@ type TracksSectionProps = {
   tracks: SimplifiedTrackObject[] | undefined;
   limit?: number;
   title?: string;
+  queueTracks?: boolean;
 };
 
-export function TracksSection({ tracks, limit, title = "Songs" }: TracksSectionProps) {
+export function TracksSection({ tracks, limit, title = "Songs", queueTracks }: TracksSectionProps) {
   if (!tracks) return null;
 
   const items = tracks.slice(0, limit || tracks.length);
 
   return (
     <List.Section title={title}>
-      {items?.map((track) => (
-        <TrackListItem
-          key={track.id}
-          track={track}
-          album={track.album}
-          playingContext={track.album?.uri}
-          showGoToAlbum
-        />
-      ))}
+      {items?.map((track) => {
+        return (
+          <TrackListItem
+            key={track.id}
+            track={track}
+            album={track.album}
+            showGoToAlbum
+            tracksToQueue={queueTracks ? tracks.filter((t) => t.id !== track.id) : undefined}
+          />
+        );
+      })}
     </List.Section>
   );
 }