+++ extensions/jira/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Jira Changelog
 
+## [Remove Raycast signature] - 2023-04-19
+
+- Remove Raycast signature preference from the `Create Issue` command
+
+## [Fix persisted values] - 2023-04-18
+
+- Fix an issue where the filter dropdown or the active sprint dropdown wouldn't save the selected dropdown value.
+
 ## [Fix priority issue] - 2023-04-13
 
 - Fix an issue where priorities could be unset, causing the loading of issues to fail.